requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
qrcode==6.1
six==1.15.0
Werkzeug==1.0.1
